// src/app/uporabnik/models/racun.ts
export class Racun {
    id: Number;
    customerId: Number;
    customerUsername: String;
    customerEmail: String;
    customerFirstName: String;
    customerLastName: String;
    polnilnicaId: Number;
    price: Number;
    status: String;
    terminDateFrom: Number;
    terminDateTo: Number;
    terminId: Number;
    timestamp: Number;
}